autoconf: change tirpc to check for a file, not for an include
Recent autoconf don't like variables in AC_CHECK_INCLUDE so we get a
warning.
In libtirpc.m4 we only need to check for the existence of a file, we
don't need to extra  support for includes, such as defining HAVE_TIRPC_H
or whatever.

So change from AC_CHECK_INCLUDE to AC_CHECK_FILE.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Steve Dickson <steved@redhat.com>