nfs-blkmapd: Fix the error status when nfs-blkmapd stops
The systemctl stop nfs-blkmap.service will sends the SIGTERM signal
to the nfs-blkmap.service first.If the process fails to be stopped,
it sends the SIGKILL signal again to kill the process.
However, exit(1) is executed in the SIGTERM processing function of
nfs-blkmap.service. As a result, systemd receives an error message
indicating that nfs-blkmap.service failed.
"Active: failed" is displayed when the systemctl status
nfs-blkmap.service command is executed.

Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: zhanchengbin <zhanchengbin1@huawei.com>
Signed-off-by: Steve Dickson <steved@redhat.com>